% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{smash_balance}
\alias{smash_balance}
\title{Balance positive and negative reviews}
\usage{
smash_balance(data_frame)
}
\arguments{
\item{data_frame}{a data frame, as produced by smash_clean()}
}
\value{
a balanced data frame (i.e. equal number of positive and negative reviews)
}
\description{
Balance positive and negative reviews
}
