% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{smash_clean}
\alias{smash_clean}
\title{Clean raw text}
\usage{
smash_clean(data_frame, parallel = TRUE)
}
\arguments{
\item{data_frame}{a data frame with a \code{raw_text} and a \code{rating} column}
}
\value{
a data frame with a new column called \code{text}
}
\description{
Clean raw text
}
