% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso.R
\name{smash_lasso}
\alias{smash_lasso}
\title{Lasso Model}
\usage{
smash_lasso(token_data_frame, original_data_frame, lambda = c("min", "1se"))
}
\arguments{
\item{token_data_frame}{tokens}

\item{original_data_frame}{origianl data frame}

\item{lambda}{the type of lambda to choose from cross-validation (default: min)}
}
\value{
A list with three elements: model coefficients, test accuracy, expanded data frame
}
\description{
Lasso Model
}
